Remove Tailwind Owl selectors
I forked this plugin to solve a specific problem but it was never used after 2019. Sometime last year I deleted the fork because I had no need of it and didn’t realize it was added to the Oakland repo without being moved *to* Oakland’s available repositories. This commit resolves that issue by removing it entirely.

This plugin was obviated by [Tailwind’s Space Between](https://tailwindcss.com/docs/space) class, which operates in the same manner.