refactor: :zap: Remove default Tailwind settings
We don’t need to declare only the exact sizes and elements we need. Rather, we should make use of `theme.extend` to add sizes, opacities, and the like to the Tailwind system.

Additionally, as there is interest in being more specific in how the `container` class works with alignment, I’ve dropped the old config setting to center it and instaed have applied `mx-auto` to all existing templates that have the `container` class in them.