feat(page layout updates): :lipstick: Update indices and entries to use max width of 1152px
Updates the following content blocks with a maximum width of `72rem` and either `2rem` or `4rem` gutters:

- Services Index
- Service Category
- Service Entry
- Public Meetings Index
- Public Meetings Entry
- Events Index
- Event Entry

It’s probably a good idea to start moving our use of flexbox layouts that have negative margins and extra padding to use the `gap` property, which is now available on all browsers. Less mental math is a good thing! 🤓