refactor(px2rem): :art: convert px measurements to tailwind classes
Goes through the existing CSS and converts all uses of `font-size` and `line-height` to their equivalent css classes in Tailwind and uses the `@apply` directive at the top of the declaration for each.

This may affect some line spacing, but most of that is from the Tailwind update in the first place. If anything text will be more legible at first and we'll want to compress the line spacing in certain areas as needed.