🗑️ (news) add portrait conditional
Adds a conditional to only include the portrait partial if an image is provided. While this update isn’t 100% necessary as we check inside the include for the portrait value as well, we *are* attempting to pass image alt text, and failing that, alt text we’ve manually generated. It should normally fail quietly, but I decided it’s easiest and least confusing if we don’t pass alt text when we are already not passing a portrait image.