🐛 (events) ensure entry is active when calling to staff
Because of how this is set up, if the contact that is selected is no longer with the City, it will return a null value but still continue querying it as though it weren't. This looks okay on production because the template fails gracefully, but locally it will cause headaches and not render properly.